# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RoadScan - Pothole & Litter Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface BoundingBox {
  /** Left edge (0-1 relative) */
  x: number;
  /** Top edge (0-1 relative) */
  y: number;
  /** Box width (0-1 relative) */
  width: number;
  /** Box height (0-1 relative) */
  height: number;
}

export type DetectedObjectClassName = typeof DetectedObjectClassName[keyof typeof DetectedObjectClassName];


export const DetectedObjectClassName = {
  pothole: 'pothole',
  plastic_waste: 'plastic_waste',
  other_litter: 'other_litter',
} as const;

export interface DetectedObject {
  id: string;
  className: DetectedObjectClassName;
  /** Confidence score 0-1 */
  confidence: number;
  bbox: BoundingBox;
  /**
     * Frame number for video detections
     * @nullable
     */
  frameNumber?: number | null;
}

export type DetectionMediaType = typeof DetectionMediaType[keyof typeof DetectionMediaType];


export const DetectionMediaType = {
  image: 'image',
  video: 'video',
} as const;

/**
 * Computed severity based on detections
 */
export type DetectionSeverity = typeof DetectionSeverity[keyof typeof DetectionSeverity];


export const DetectionSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface DetectionCounts {
  pothole: number;
  plastic_waste: number;
  other_litter: number;
  total: number;
}

export interface Detection {
  id: number;
  filename: string;
  mediaType: DetectionMediaType;
  /** URL to the original uploaded file */
  originalUrl: string;
  /** URL to the annotated output file */
  annotatedUrl: string;
  /**
     * URL to a small thumbnail
     * @nullable
     */
  thumbnailUrl?: string | null;
  objects: DetectedObject[];
  counts: DetectionCounts;
  processingTimeMs: number;
  /** Computed severity based on detections */
  severity: DetectionSeverity;
  /**
     * Optional location label
     * @nullable
     */
  location?: string | null;
  /**
     * GPS latitude from EXIF or manual input
     * @nullable
     */
  lat?: number | null;
  /**
     * GPS longitude from EXIF or manual input
     * @nullable
     */
  lon?: number | null;
  createdAt: string;
}

export interface DetectionListResponse {
  items: Detection[];
  total: number;
}

export type StatsResponseSeverityBreakdown = {
  low: number;
  medium: number;
  high: number;
  critical: number;
};

export interface DailyActivity {
  date: string;
  scans: number;
  objects: number;
}

export interface StatsResponse {
  totalScans: number;
  totalObjects: number;
  classBreakdown: DetectionCounts;
  severityBreakdown: StatsResponseSeverityBreakdown;
  avgConfidence: number;
  avgProcessingMs: number;
  recentActivity: DailyActivity[];
}

export type ListDetectionsParams = {
limit?: number;
offset?: number;
mediaType?: ListDetectionsMediaType;
};

export type ListDetectionsMediaType = typeof ListDetectionsMediaType[keyof typeof ListDetectionsMediaType];


export const ListDetectionsMediaType = {
  image: 'image',
  video: 'video',
  all: 'all',
} as const;