# lib/api-zod/src/generated/types/dailyActivity.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RoadScan - Pothole & Litter Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyActivity {
  date: string;
  scans: number;
  objects: number;
}