# lib/api-zod/src/generated/types/detection.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RoadScan - Pothole & Litter Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { DetectedObject } from './detectedObject';
import type { DetectionCounts } from './detectionCounts';
import type { DetectionMediaType } from './detectionMediaType';
import type { DetectionSeverity } from './detectionSeverity';

export interface Detection {
  id: number;
  filename: string;
  mediaType: DetectionMediaType;
  /** URL to the original uploaded file */
  originalUrl: string;
  /** URL to the annotated output file */
  annotatedUrl: string;
  /**
     * URL to a small thumbnail
     * @nullable
     */
  thumbnailUrl?: string | null;
  objects: DetectedObject[];
  counts: DetectionCounts;
  processingTimeMs: number;
  /** Computed severity based on detections */
  severity: DetectionSeverity;
  /**
     * Optional location label
     * @nullable
     */
  location?: string | null;
  /**
     * GPS latitude from EXIF or manual input
     * @nullable
     */
  lat?: number | null;
  /**
     * GPS longitude from EXIF or manual input
     * @nullable
     */
  lon?: number | null;
  createdAt: Date;
}