# lib/api-zod/src/generated/types/detectionListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RoadScan - Pothole & Litter Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { Detection } from './detection';

export interface DetectionListResponse {
  items: Detection[];
  total: number;
}